feat : change exchange response form
교환 관련 API response 구조 수정
